+++ release.sh
@@ -1,11 +1,33 @@
 #! /usr/bin/bash
 set -uvx
 set -e
-pubspec-gen
+cwd=`pwd`
+ts=`date "+%Y.%m%d.%H%M"`
+version="${ts//.0/.}"
+
+echo $version
+
+pubspec-gen "$version"
 #dart pub get
+
+cat << EOS >> CHANGELOG.md
+
+## $version
+
+- pubspec.yaml was generated using pubspec-gen
+EOS
+
+
 ./do-analyze.sh
 ./do-test.sh
-dart pub publish
+
+tag="$version"
+#cd $cwd
 git add .
-git commit -m.
-git push
+git commit -m"$tag"
+git tag -a "$tag" -m"$tag"
+git push origin "$tag"
+git push origin HEAD:main
+git remote -v
+
+#dart pub publish