+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+## 0.0.1
+
+- Initial release
+
 ## 0.0.2
 
 - Modified dart_scan.dart so that scan exclude comments
 
-## 0.0.1
+## 2025.420.754
 
-- Initial release
+- pubspec.yaml was generated using pubspec-gen
+
+## 2025.420.756
+
+- pubspec.yaml was generated using pubspec-gen